APP/requirements.txt
gradio>=4.0.0
streamlit>=1.30.0
tensorflow>=2.10.0
numpy>=1.22.0
Pillow>=9.0.0
opencv-python-headless>=4.5.5
keras